requirements.txt
aiogram==3.13.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis==5.2.1
apscheduler==3.10.4
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.12
python-dotenv==1.0.1
